move trace.hpp
