+++ tt_metal/distributed/mesh_trace.hpp
@@ -6,7 +6,7 @@
 
 #include <tt-metalium/mesh_buffer.hpp>
 #include <tt-metalium/mesh_trace_id.hpp>
-#include <tt-metalium/trace_buffer.hpp>
+#include "trace/trace_buffer.hpp"
 
 namespace tt::tt_metal::distributed {
