+++ tt_metal/impl/dispatch/hardware_command_queue.cpp
@@ -9,7 +9,7 @@
 // Because we are a Friend of Program, accessing Program::get_program_transfer_info() and Program::get_kernels_buffer()
 // MUST REMOVE
 #include <tt-metalium/program.hpp>
-#include <trace_buffer.hpp>
+#include "trace/trace_buffer.hpp"
 #include <tracy/Tracy.hpp>
 #include <tt-metalium/allocator.hpp>
 #include <tt_stl/overloaded.hpp>
