+++ tt_metal/impl/dispatch/hardware_command_queue.hpp
@@ -26,7 +26,7 @@
 #include "tt-metalium/program.hpp"
 #include <tt_stl/span.hpp>
 #include "sub_device_types.hpp"
-#include "trace_buffer.hpp"
+#include "trace/trace_buffer.hpp"
 #include "tt_metal/impl/buffers/dispatch.hpp"
 #include <umd/device/tt_core_coordinates.h>
 #include "vector_aligned.hpp"
