+++ tt_metal/impl/dispatch/host_runtime_commands.hpp
@@ -23,7 +23,7 @@
 #include "dispatch_settings.hpp"
 #include "tt-metalium/program.hpp"
 #include "sub_device_types.hpp"
-#include "trace_buffer.hpp"
+#include "trace/trace_buffer.hpp"
 #include "tt_metal/impl/program/program_command_sequence.hpp"
 #include "worker_config_buffer.hpp"
