+++ tt_metal/impl/lightmetal/lightmetal_capture.cpp
@@ -8,7 +8,7 @@
 #include "flatbuffers/flatbuffers.h"
 #include "command_generated.h"
 #include "light_metal_binary_generated.h"
-#include <trace_buffer.hpp>
+#include "trace/trace_buffer.hpp"
 #include <tt-metalium/buffer.hpp>
 #include <tt-metalium/program.hpp>
 #include <tt-metalium/kernel.hpp>
