+++ tt_metal/impl/trace/dispatch.hpp
@@ -6,7 +6,6 @@
 
 #include <device.hpp>
 #include <stdint.h>
-#include <trace_buffer.hpp>
 #include <cstddef>
 #include <unordered_map>
 #include <vector>
@@ -15,6 +14,7 @@
 #include "dispatch/dispatch_settings.hpp"
 #include "dispatch/worker_config_buffer.hpp"
 #include "sub_device_types.hpp"
+#include "trace_buffer.hpp"
 
 namespace tt {
 namespace tt_metal {
